Correction to RAPL information
this is happening because the energy status MSR registers for Package: "MSR_PKG_ENERGY_STATUS" and DRAM: "MSR_DRAM_ENERGY_STATUS"
are counted on 32 bit ( the MSRs are 64bit wide) and they wrap around every few minutes.

for info check out the attached screenshot of the Intel's doc
or here for Intel's complete doc: http://download.intel.com/products/processor/manual/325384.pdf

+++ src/plugins/acct_gather_energy/rapl/acct_gather_energy_rapl.c
@@ -70,8 +70,8 @@
 union {
 	uint64_t val;
 	struct {
-		uint64_t low;
-		uint64_t high;
+		uint32_t low;
+		uint32_t high;
 	} i;
 } package_energy[MAX_PKGS], dram_energy[MAX_PKGS];
